refactor: MovieDetailViewController Cell 분리
